updated history output
Created Hist1Mod.F90 to write history records.
No answer change is expected.

+++ f90src/ecosim_datatype/FlagDataType.F90
@@ -22,7 +22,7 @@ implicit none
   integer :: IWTHR(2)              !weather data type:1=daily,2=hourly for first(L=1) or second(L=2) scene
   integer :: IUTYP(JY,JX)          !urea hydrolysis inhibitor type (1=no,2=yes)
   integer :: IYTYP(0:2,366,JY,JX)  !fertilizer release type from fertilizer input file
-  real(r8) :: ITILL(366,JY,JX)     !soil disturbance type, [-]
+  integer :: ITILL(366,JY,JX)      !soil disturbance type, [-]
 
   integer :: IFLGC(JP,JY,JX)       ! flag for living pft
   integer :: IFLGI(JP,JY,JX)       !PFT initialization flag:0=no,1=yes
@@ -36,7 +36,7 @@ implicit none
   integer :: IBTYP(JP,JY,JX)       !phenologically-driven above-ground turnover (all, foliar only, none)
   integer :: IRTYP(JP,JY,JX)       !grain type (below or above-ground)
   integer :: MY(JP,JY,JX)          !mycorrhizal type (no or yes)
-  real(r8) :: IDTBL(JY,JX)         !water table flag from site file
+  integer :: IDTBL(JY,JX)         !water table flag from site file
 
 
 end module FlagDataType
